@@ -6,14 +6,14 @@ publish_to: 'none'
 version: 1.1.0
 
 environment:
-  sdk: ">=2.18.0 <3.0.0"
-  flutter: 3.3.10
+  sdk: ">=2.19.2 <3.0.0"
+  flutter: 3.7.5
 
 dependencies:
   flutter:
     sdk: flutter
-  flutter_hooks: ^0.18.5+1
-  hooks_riverpod: ^2.1.1
+  flutter_hooks: ^0.18.6
+  hooks_riverpod: ^2.2.0
   flutter_localizations:
     sdk: flutter
   linked_scroll_controller: ^0.2.0
@@ -21,41 +21,39 @@ dependencies:
   highlight: ^0.7.0
   flutter_highlight: ^0.7.0
   resizable_widget: ^1.0.5
-  multi_split_view: ^2.1.0
-  dio: ^4.0.6
-  dropdown_search: ^3.0.1
-  shared_preferences: ^2.0.13
-  drift: ^1.7.0
-  sqlite3_flutter_libs: ^0.5.12
-  path_provider: ^2.0.0
-  path: ^1.8.1
-  hive: ^2.2.2
+  multi_split_view: ^2.4.0
+  dio: ^5.0.1
+  dropdown_search: ^5.0.6
+  shared_preferences: ^2.0.18
+  sqlite3_flutter_libs: ^0.5.13
+  path_provider: ^2.0.13
+  path: ^1.8.2
+  hive: ^2.2.3
   hive_flutter: ^1.1.0
-  freezed_annotation: ^2.0.3
+  freezed_annotation: ^2.2.0
   bitsdojo_window: ^0.1.5
-  connectivity_plus: ^2.3.9
-  package_info_plus: ^1.4.2
+  connectivity_plus: ^3.0.3
+  package_info_plus: ^3.0.3
   clipboard: ^0.1.3
-  fpdart: ^0.2.0
-  supabase_flutter: ^0.3.3
-  desktop_drop: ^0.4.0
-  flutter_downloader: ^1.9.1
-  flutter_svg: ^1.1.6
-  test: ^1.21.4
+  fpdart: ^0.4.1
+  supabase_flutter: ^1.4.1
+  desktop_drop: ^0.4.1
+  flutter_downloader: ^1.10.2
+  flutter_svg: ^2.0.2
+  test: ^1.22.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  drift_dev: ^1.7.0
   build_runner: ^2.1.11
 
   flutter_lints: ^1.0.0
-  freezed: ^2.2.0
-  json_serializable: ^6.5.4
-  json_annotation: ^4.7.0
-  flutter_launcher_icons: "^0.11.0"
-  msix: ^3.7.0
+  freezed: ^2.3.2
+  json_serializable: ^6.6.1
+  json_annotation: ^4.8.0
+  flutter_launcher_icons: ^0.12.0
+  msix: ^3.8.2
 
 flutter:
   uses-material-design: true
