@@ -5,16 +5,18 @@
 import FlutterMacOS
 import Foundation
 
+import app_links
 import bitsdojo_window_macos
-import connectivity_plus_macos
+import connectivity_plus
 import desktop_drop
-import package_info_plus_macos
-import path_provider_macos
-import shared_preferences_macos
+import package_info_plus
+import path_provider_foundation
+import shared_preferences_foundation
 import sqlite3_flutter_libs
 import url_launcher_macos
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
+  AppLinksMacosPlugin.register(with: registry.registrar(forPlugin: "AppLinksMacosPlugin"))
   BitsdojoWindowPlugin.register(with: registry.registrar(forPlugin: "BitsdojoWindowPlugin"))
   ConnectivityPlugin.register(with: registry.registrar(forPlugin: "ConnectivityPlugin"))
   DesktopDropPlugin.register(with: registry.registrar(forPlugin: "DesktopDropPlugin"))
